@@ -15,7 +15,7 @@
     "turbo": "^1.9.3"
   },
   "engines": {
-    "node": ">=16.0.0"
+    "node": ">=18.0.0"
   },
   "workspaces": [
     "packages/*",
