@@ -8,7 +8,7 @@ type Props = {
 };
 
 export const generators: { [key: string]: string; } = {
-  Wordpress: 'https://wordpress.org/',
+  WordPress: 'https://wordpress.org/',
   Ghost: 'https://ghost.org/',
   Jekyll: 'https://jekyllrb.com/',
   Hugo: 'https://gohugo.io/',
@@ -42,8 +42,6 @@ export const feedFormats: { [key: string]: string; } = {
 };
 
 export const Blog: React.FunctionComponent<Props> = ({ blog }) => {
-  const generator = blog.generator ? blog.generator.split(' ')[0] : null;
-
   return (
     <div className="bg-inherit pt-4">
       <div className="mx-auto max-w-7xl px-6 lg:px-8">
@@ -76,7 +74,7 @@ export const Blog: React.FunctionComponent<Props> = ({ blog }) => {
           <div className="mt-2">
             <span className="text-gray-500">
               <Link
-                href={blog.homePageUrl ?? ''}
+                href={blog.homepage_url ?? ''}
                 target="_blank"
                 className="relative mr-6 w-0 py-2 text-base font-medium"
               >
@@ -84,33 +82,23 @@ export const Blog: React.FunctionComponent<Props> = ({ blog }) => {
                 <span className="ml-2">Home Page</span>
               </Link>
             </span>
-            {blog.feedUrl && blog.feedFormat && (
+            {blog.feed_url && blog.feed_format && (
               <span className="-ml-px text-gray-500">
-                <Link href={blog.feedUrl} target="_blank" className="relative mr-6 w-0 py-2 text-base font-medium">
+                <Link href={blog.feed_url} target="_blank" className="relative mr-6 w-0 py-2 text-base font-medium">
                   <Icon icon="fa6-solid:rss" className="inline" />
-                  <span className="ml-2">{feedFormats[blog.feedFormat] + ' Feed'}</span>
+                  <span className="ml-2">{feedFormats[blog.feed_format] + ' Feed'}</span>
                 </Link>
               </span>
             )}
-            <span className="-ml-px text-gray-500">
-              <Link
-                href={'/' + blog.id + '.json'}
-                target="_blank"
-                className="relative mr-6 w-0 py-2 text-base font-medium"
-              >
-                <Icon icon="fa6-solid:square-rss" className="inline" />
-                <span className="ml-2">JSON Feed</span>
-              </Link>
-            </span>
-            {generator && (
+            {blog.generator && (
               <span className="text-gray-500">
                 <Link
-                  href={generators[generator] ?? ''}
+                  href={generators[blog.generator] ?? ''}
                   target="_blank"
                   className="relative mr-6 py-2 text-base font-medium text-gray-500"
                 >
                   <Icon icon="fa6-solid:rocket" className="inline" />
-                  <span className="ml-2">{generator}</span>
+                  <span className="ml-2">{blog.generator}</span>
                 </Link>
               </span>
             )}
@@ -133,11 +121,11 @@ export const Blog: React.FunctionComponent<Props> = ({ blog }) => {
                 <span className="ml-2 mr-6">License not confirmed</span>
               </span>
             )}
-            {blog.dateIndexed && (
+            {blog.indexed_at && (
               <span className="text-gray-500">
                 <Icon icon="fa6-regular:calendar-plus" className="inline" />
-                <time className="ml-2 mr-6" dateTime={blog.dateIndexed.toString()}>
-                  {new Date(blog.dateIndexed).toLocaleDateString('en-US', {
+                <time className="ml-2 mr-6" dateTime={blog.indexed_at.toString()}>
+                  {new Date(blog.indexed_at).toLocaleDateString('en-US', {
                     year: 'numeric',
                     month: 'long',
                     day: 'numeric',
@@ -145,7 +133,7 @@ export const Blog: React.FunctionComponent<Props> = ({ blog }) => {
                 </time>
               </span>
             )}
-            {!blog.dateIndexed && (
+            {!blog.indexed_at && (
               <span className="text-orange-600">
                 <Icon icon="fa6-solid:eye-slash" className="inline" />
                 <span className="ml-2 mr-6">Preview only</span>
