@@ -72,7 +72,7 @@ export const Posts: React.FunctionComponent<Props> = ({ posts }) => {
                         <Link href={post.url}>{post.title}</Link>
                       </h3>
                     )}
-                    <Byline authors={post.authors} datePublished={post.datePublished} />
+                    <Byline authors={post.authors} datePublished={post.published_at} />
                     {post.summary && (
                       <p className="text-medium mt-2 leading-6 text-gray-900">{parse(String(post.summary))}</p>
                     )}
