@@ -1,3 +1,3 @@
 // Define the path to which the user will be redirected after login / signup
-export const redirectPath = '/app';
+export const redirectPath = '/account';
 export const accountPath = '/account';
