@@ -1,5 +1,4 @@
 import { omit } from 'lodash';
-import { GetStaticPaths } from 'next';
 import Head from 'next/head';
 import Link from 'next/link';
 import React from 'react';
@@ -7,27 +6,17 @@ import { jsonLdScriptProps } from 'react-schemaorg';
 import { Blog as BlogSchema } from 'schema-dts';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 
-import { getAllConfigs } from '@/pages/api/blogs';
-import { getSingleBlog } from '@/pages/api/blogs/[slug]';
+import { getSingleBlog, getPosts } from '@/pages/api/blogs/[slug]';
 import { BlogType, PostType } from '@/types/blog';
 import Layout from '@/components/layout/Layout';
 import { Blog } from '@/components/common/Blog';
 import { Posts } from '@/components/common/Posts';
 
-export const getStaticPaths: GetStaticPaths = async () => {
-  const configs = await getAllConfigs();
-  const paths = configs.map((config) => ({
-    params: { slug: config.id },
-  }));
-
-  return { paths, fallback: false };
-};
-
-export async function getStaticProps({ params }) {
-  const blog = await getSingleBlog(params.slug, { includePosts: true });
-
-  return { props: { ...(await serverSideTranslations('en', ['common', 'app'])), blog, posts: blog.items } };
-};
+export async function getServerSideProps(ctx) {
+  const blog = await getSingleBlog(ctx.params.slug);
+  const posts = await getPosts(ctx.params.slug);
+  return { props: { ...(await serverSideTranslations('en', ['common', 'app'])), blog, posts } };
+}
 
 type Props = {
   blog: BlogType;
@@ -64,14 +53,14 @@ const BlogPage: React.FunctionComponent<Props> = ({ blog, posts }) => {
         />
       </Head>
       <Layout>
-        <div className={blog.dateIndexed ? 'bg-white' : 'bg-blue-50'}>
+        <div className={blog.indexed_at ? 'bg-white' : 'bg-blue-50'}>
           <Blog blog={blog} />
-          <Posts posts={posts} />
-          {blog.homePageUrl && (
+          {posts && <Posts posts={posts} />}
+          {blog.homepage_url && (
             <div className="mx-auto max-w-2xl bg-inherit pb-2 lg:max-w-4xl">
               <div className="my-5 lg:my-8">
                 <Link
-                  href={blog.homePageUrl}
+                  href={blog.homepage_url}
                   target="_blank"
                   className="text-xl font-semibold text-gray-700 hover:text-gray-400 sm:text-xl"
                 >
