@@ -11,13 +11,13 @@ export interface PostType {
   url?: string;
   title?: string;
   summary?: string;
-  datePublished?: string;
-  dateModified?: string;
+  published_at?: string;
+  modified_at?: string;
   authors?: AuthorType[];
   image?: string;
   thumbnail?: string;
-  contentHtml?: string;
-  contentText?: string;
+  content_html?: string;
+  content_text?: string;
   tags?: string[];
   language?: string;
   issn?: string;
@@ -30,17 +30,17 @@ export interface BlogType extends Omit<FeedData, 'entries' | 'published' | 'link
   category?: string;
   description?: string;
   language?: string;
-  baseUrl?: string;
-  homePageUrl?: string;
-  feedUrl?: string;
+  base_url?: string;
+  homepage_url?: string;
+  feed_url?: string;
   icon?: string;
   favicon?: string;
-  dateModified?: string;
-  dateIndexed?: string;
+  modified_at?: string;
+  indexed_at?: string;
   generator?: string;
-  hasLicense?: boolean;
+  has_license?: boolean;
   license?: string;
-  feedFormat?: string;
+  feed_format?: string;
   items?: PostType[];
   expired?: boolean;
 }
